third_party/WebKit/LayoutTests/fast/css/namespaces/default-namespace-insert-rule.html

<!doctype html>

<style>
@namespace url(banana);
*|div {
    width: 100px;
    height: 100px;
    background-color: green;
}
</style>
<div id=target></div>

<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>

<script>
test(function(){
    document.styleSheets[0].insertRule('#target { background-color: red; }', 1);
    assert_equals(getComputedStyle(target).backgroundColor, 'rgb(0, 128, 0)');
}, "CSSStyleSheet::insertRule should respect default namespaces");
</script>
